Update of modalMor (Issue #13)
+ Input changed to Opts

+ Splitting of complex conjugated columns of V:
In the old code all columns of V have been splitted into real and
imaginary ones (-> only even output order), even if they have been
purely real (-> imaginary column containing only zeros and successive
(real) column completely ignored/lost) with no previous sorting of the
columns (eigs returns the eigenvalues unsorted, so successive columns
don't have to be complex conjugated)
-> Sorting of the eigenvalues, so that only truly complex conjugated
columns are splitted
-> No splitting of purely real columns

+ Calculation of the left eigenvector:
the input 'sigma' to eigs has been the eigenvalue of A -> eigs returns
an error if sigma is an eigenvalue (e.g. if A=A')
-> sigma = eigenvalue - 1e-6

+ Only even reduction orders possible:
All reduction orders are possible, but sometimes the obtained reduced
system is then complex:
E.g. if the eigenvalues of A are =[-1,-2+i,-2-i,-5] with 'SM' and q=2
(even) is chosen:
-> eigs(A,q,'SM') returns the eigenvalues = [-1, -2+i]  -> just one
complex column without partner in V
-> splitting in real and imaginary column not possible
-> complex reduced system (despite the even order).
-> warning in code