Create filter for doc/html/helptoc.xml
The "doc/html/helptoc.xml file" is marked with a timestamp
every time it is generated by the publishHelp.m function,
even if the rest of its contents haven't changed. This
causes a versioning problem, because git constantly detects
this as a change in the file.

In order to avoid this problem, a filter named
"timestamp_reset" was created in the "helptoc_filter" file.
The filter was then assigned to the "doc/html/helptoc.xml" file
by the ".gitattributes" file.

IMPORTANT: in order for the filter to work, the user has to
manually change the ".git/config" file in the repository
(it is a hidden file).
To do this, type in the Git shell:

git config --local include.path "absolute/path/to/helptoc_filter"

Alternatively you can manually add the following lines to the
".git/config" file:

[include]
     path = absolute/path/to/helptoc_filter