Add header support for classdef files
The function funHelp2Markup was designed to parse the headers of
MATLAB functions. The way it determines if an M-file is a function
is by parsing the very first line in the file.
Before this update it would only capture the function name if the
first line had the word "function" at the beggining. With this
update it is now it is also able to capture the name of classdef files.
This allows now for the parsing of the header of e.g. the sss classdef
file.

The functions_index.m and helptoc.xml files were also updated now that
the sss classdef header is included in the documentation.