HinfMor - using 3D arrays for DrInit and checking stability
For the Dr sweeps 3D arrays are used.
Further, it the initialization type is not set to 0, then stability of the model initialized ad Dr0 is tested. Since DrInit can call itself sometimes, this operation is conducted only for the first execution of DrInit (if it is not calling itself).