testFreqresp: added tests for new functionality
freqresp is now aligned with the builtin case and accepts real, imaginary and complex frequencies.
If the frequencies are all real, then the evaluation occurs on the imaginary axis and a real vector is returned, otherwise the evaluation is at the frequencies specified by the user and the vector returned is exactly the same that was passed.

NOTE: this is aligned with the built-in version. However, if you want to evaluate the transfer function over the real axis then this implementation would fail. Evaluate manually or add an imaginary frequency to resolve