App: bugfix concerning header and footer
Bug at the functionality that does not display header and footer if the screen is to small. Instead of the real current size of the app window, the desired size has to be taken for the calculation of the rescaling of header and footer. The reason for this is that MATLAB automatically decreases the size of the app window if the screen is to small. Therefore, the real app window size can be unequal to the desired App window size